requirements.txt
colorama
lxml
pycrypto
